+++ src/driver_select.h
@@ -8,6 +8,16 @@
 
 #include "qnethernet_opts.h"
 
+// First check for the existence of an external driver
+// https://forum.pjrc.com/index.php?threads/new-lwip-based-ethernet-library-for-teensy-4-1.68066/post-345539
+#if defined(__has_include)
+// https://gcc.gnu.org/onlinedocs/cpp/_005f_005fhas_005finclude.html
+#if __has_include(<qnethernet_external_driver.h>)
+#define QNETHERNET_INTERNAL_DRIVER_EXTERNAL
+#endif  // __has_include(<qnethernet_external_driver.h>)
+#endif  // defined(__has_include)
+
+#ifndef QNETHERNET_INTERNAL_DRIVER_EXTERNAL
 #if defined(QNETHERNET_DRIVER_W5500)
 #include "drivers/driver_w5500.h"
 #define QNETHERNET_INTERNAL_DRIVER_W5500
@@ -18,3 +28,4 @@
 #include "drivers/driver_unsupported.h"
 #define QNETHERNET_INTERNAL_DRIVER_UNSUPPORTED
 #endif  // Driver selection
+#endif  // QNETHERNET_INTERNAL_DRIVER_EXTERNAL