+++ src/lwip_driver.h
@@ -47,6 +47,16 @@
 //    enum value. If no change is made, the default 'EthernetOtherHardware' will
 //    be returned if hardware is found (driver_has_hardware() returns true).
 
+// How to create an external driver that isn't inside the distribution:
+// 1. Create a header named "qnethernet_external_driver.h" that defines MTU and
+//    MAX_FRAME_LEN. Don't forget to use either `#pragma once` or a
+//    #define guard.
+// 2. Add lwIP options (see lwip/opt.h) with appropriate values for your driver.
+//    For example, Ethernet padding, checksum generation, and checksum checking.
+// 3. Create driver source and include lwip_driver.h. Implement all the
+//    `driver_x()` functions. It can be written in either C or C++. If C++ then
+//    make sure to use `extern "C"` around those functions.
+
 #ifdef __cplusplus
 extern "C" {
 #endif  // __cplusplus